refactor: rewrite domain design
- change the folders structure
- normalize indentation of yaml files (2 spaces)
