+++ src/Domain/Common/Model/Exception/StringExceeds255Characters.php
@@ -11,10 +11,10 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Domain\Model\Supplier;
+namespace Domain\Common\Model\Exception;
 
-use Domain\Common\Model\ContactUuid;
-
-class SupplierUuid extends ContactUuid
+final class StringExceeds255Characters extends \DomainException
 {
+    /** @var string */
+    protected $message = 'Le titre est trop long. Vous ne devriez pas exéder 255 caractères';
 }
