--- src/Domain/Model/Common/VO/EmailField.php
@@ -1,39 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of the  G.L.S.R. Apps package.
- *
- * (c) Dev-Int Création <info@developpement-interessant.com>.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Domain\Model\Common\VO;
-
-use Domain\Model\Common\Exception\InvalidEmail;
-
-final class EmailField
-{
-    private string $email;
-
-    public function __construct(string $email)
-    {
-        if (!\filter_var($email, \FILTER_VALIDATE_EMAIL)) {
-            throw new InvalidEmail();
-        }
-        $this->email = $email;
-    }
-
-    public static function fromString(string $email): self
-    {
-        return new self($email);
-    }
-
-    public function getValue(): string
-    {
-        return $this->email;
-    }
-}
