+++ tests/Unit/Domain/Common/Model/VO/EmailFieldTest.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the  G.L.S.R. Apps package.
+ *
+ * (c) Dev-Int Création <info@developpement-interessant.com>.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Unit\Tests\Domain\Common\Model\VO;
+
+use Domain\Common\Model\Exception\InvalidEmail;
+use Domain\Common\Model\VO\EmailField;
+use PHPUnit\Framework\TestCase;
+
+class EmailFieldTest extends TestCase
+{
+    final public function testCreateWithInvalidEmailThrowsADomainException(): void
+    {
+        // Arrange
+        $this->expectException(InvalidEmail::class);
+
+        // Act & Assert
+        EmailField::fromString('invalid.email.fr');
+    }
+}
